requirements.txt
ipython==3.2.1
plop==0.1.1
tornado==4.2.1
sockjs-tornado==1.0.1
tornado-redis==2.4.18